% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pep_codebook.R
\name{get_pep_codebook}
\alias{get_pep_codebook}
\title{Get the cori.data.pep variable codebook}
\usage{
get_pep_codebook()
}
\value{
A data frame with columns: \code{variable}, \code{label}, \code{unit}, \code{nominal},
\code{notes}.
}
\description{
Returns documentation for all variables produced by the CORI PEP processing
pipeline.
}
\examples{
get_pep_codebook()

}
\seealso{
\code{\link[=read_pep_from_s3]{read_pep_from_s3()}}
}
