% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pep.R
\name{pull_components}
\alias{pull_components}
\title{Pull components of population change across multiple years}
\usage{
pull_components(years = 2000:as.integer(format(Sys.Date(), "\%Y")))
}
\arguments{
\item{years}{Integer vector. Years to return. Default: \code{2000} to current year.}
}
\description{
Returns births, deaths, natural change, domestic migration, international
migration, and net migration for all U.S. counties, 2000 to present.
}
\keyword{internal}
